third_party/WebKit/LayoutTests/http/tests/inspector/network/json-preview.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
async function test()
{
    function createNetworkRequestWithJSONMIMEType(type)
    {
        InspectorTest.addResult("Creating a NetworkRequest with type: " + type);
        var request = new SDK.NetworkRequest(SDK.targetManager.mainTarget(), 0, 'http://localhost');
        request.mimeType = type;
        request._content = '{"number": 42}';
        return request;
    }

    async function testPreviewer(request)
    {
        var previewView = new Network.RequestPreviewView(request, null);
        var previewer = await previewView._createPreviewView();

        InspectorTest.addResult("Its previewer is searchable: " + (previewer && previewer instanceof UI.SearchableView));
        InspectorTest.addResult("Its previewer is the JSON previewer: " + (previewer && previewer._searchProvider && previewer._searchProvider instanceof Network.JSONView));
    }

    await testPreviewer(createNetworkRequestWithJSONMIMEType("application/json"));
    await testPreviewer(createNetworkRequestWithJSONMIMEType("application/vnd.document+json"));

    InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests that resources with JSON MIME types are previewed with the JSON viewer.</p>
</body>
</html>
