third_party/WebKit/LayoutTests/http/tests/inspector/network/json-preview-expected.txt

Tests that resources with JSON MIME types are previewed with the JSON viewer.

Creating a NetworkRequest with type: application/json
Its previewer is searchable: true
Its previewer is the JSON previewer: true
Creating a NetworkRequest with type: application/vnd.document+json
Its previewer is searchable: true
Its previewer is the JSON previewer: true
